Implemented 'PARSE' option for Assignment 1
Also updated ReadMe to describe how to launch a shell in the
Docker container, just in case someone wants to run more
specific tests.

Also changed the order of instructions in the dockerfile so
moving our code in happens last (logic being that our code is
the most likely to change, and steps that need to be re-done
more often fit better towards the end of the build).