Implemented 'Read From File Section' in Assignment 3
Notably, the correct approach would be to extract common SF-file-related
functionality between Assignment 1 and Assignment 3 into a library or such,
then have both use that (to avoid code duplication). But because the tester
needs to compile the code itself, adding external libraries without making
changes to the tester is something I'm not sure how to do cleanly.